+++ src/config.js
@@ -7,7 +7,12 @@ export const CHECK_INTERVAL = 1000 * 3600; // How often should we perform checks
 export const LOCAL_API = 'http://localhost:3000';
 export const DEV_API = 'https://dev.franzinfra.com';
 export const LIVE_API = 'https://api.franzinfra.com';
-export const GA_ID = 'UA-74126766-6';
+
+export const APP_THEMES = {
+  'theme-regular': 'Regular',
+  'theme-dark': 'Dark',
+  'theme-transparent-dark': 'Transparent dark',
+};
 
 export const DEFAULT_APP_SETTINGS = {
   autoLaunchInBackground: false,
@@ -16,6 +21,8 @@ export const DEFAULT_APP_SETTINGS = {
   minimizeToSystemTray: false,
   showDisabledServices: true,
   showMessageBadgeWhenMuted: true,
+  theme: 'theme-regular',
+  appBackground: './assets/images/galaxy2.jpg',
   enableSpellchecking: true,
   locale: '',
   fallbackLocale: 'en-US',