+++ src/containers/auth/PricingScreen.js
@@ -6,7 +6,6 @@ import { RouterStore } from 'mobx-react-router';
 import Pricing from '../../components/auth/Pricing';
 import UserStore from '../../stores/UserStore';
 import PaymentStore from '../../stores/PaymentStore';
-import { gaPage } from '../../lib/analytics';
 
 import { globalError as globalErrorPropType } from '../../prop-types';
 
@@ -15,10 +14,6 @@ export default class PricingScreen extends Component {
   static propTypes = {
     error: globalErrorPropType.isRequired,
   };
-
-  componentDidMount() {
-    gaPage('Auth/Pricing');
-  }
 
   render() {
     const { actions, stores, error } = this.props;