+++ src/containers/settings/ServicesScreen.js
@@ -6,15 +6,11 @@ import { RouterStore } from 'mobx-react-router';
 // import RecipePreviewsStore from '../../stores/RecipePreviewsStore';
 import UserStore from '../../stores/UserStore';
 import ServiceStore from '../../stores/ServicesStore';
-import { gaPage } from '../../lib/analytics';
 
 import ServicesDashboard from '../../components/settings/services/ServicesDashboard';
 
 @inject('stores', 'actions') @observer
 export default class ServicesScreen extends Component {
-  componentDidMount() {
-    gaPage('Settings/Service Dashboard');
-  }
 
   componentWillUnmount() {
     this.props.actions.service.resetFilter();