+++ src/stores/UserStore.js
@@ -7,7 +7,6 @@ import { isDevMode } from '../environment';
 import Store from './lib/Store';
 import Request from './lib/Request';
 import CachedRequest from './lib/CachedRequest';
-import { gaEvent } from '../lib/analytics';
 
 const debug = require('debug')('UserStore');
 
@@ -137,8 +136,6 @@ export default class UserStore extends Store {
     this._setUserData(authToken);
 
     this.stores.router.push('/');
-
-    gaEvent('User', 'login');
   }
 
   @action async _signup({ firstname, lastname, email, password, accountType, company }) {
@@ -156,8 +153,6 @@ export default class UserStore extends Store {
     this._setUserData(authToken);
 
     this.stores.router.push(this.PRICING_ROUTE);
-
-    gaEvent('User', 'signup');
   }
 
   @action async _retrievePassword({ email }) {
@@ -165,8 +160,6 @@ export default class UserStore extends Store {
 
     await request._promise;
     this.actionStatus = request.result.status || [];
-
-    gaEvent('User', 'retrievePassword');
   }
 
   @action async _invite({ invites }) {
@@ -180,8 +173,6 @@ export default class UserStore extends Store {
     if (this.stores.router.location.pathname.includes(this.INVITE_ROUTE)) {
       this.stores.router.push('/');
     }
-
-    gaEvent('User', 'inviteUsers');
   }
 
   @action async _update({ userData }) {
@@ -189,8 +180,6 @@ export default class UserStore extends Store {
 
     this.getUserInfoRequest.patch(() => response.data);
     this.actionStatus = response.status || [];
-
-    gaEvent('User', 'update');
   }
 
   @action _resetStatus() {